Detect login page to improve error message
Detect whether the page is a login page. Sometimes, we attempt to log in but the modem still returns the Login page. Instead of logging “Failed to parse modem datetime. HTML scrape error?” and “Failed to parse product_info/uptime as timedelta. Scrape issue?”, log an error that the Login page was returned.